@@ -32,61 +32,60 @@ import {
 import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
 
 @NgModule({
-  imports: [
-    MatFormFieldModule,
-    MatRippleModule,
-    MatDialogModule,
-    MatSortModule,
-    MatPaginatorModule,
-    CdkTableModule,
-    DragDropModule,
-    PortalModule,
-    CommonModule,
-    MatSelectModule,
-    MatCheckboxModule,
-    MatInputModule,
-    MatRadioModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatTabsModule,
-    MatButtonModule,
-    MatTableModule,
-    MatSlideToggleModule,
-    OverlayModule,
-    MatTooltipModule,
-    MatProgressBarModule,
-    MatMenuModule,
-    MatProgressSpinnerModule
-  ],
-  declarations: [],
-  entryComponents: [],
-  providers: [
-    { provide: MAT_RIPPLE_GLOBAL_OPTIONS, useValue: globalRippleConfig }
-  ],
-  exports: [
-    MatFormFieldModule,
-    PlatformModule,
-    MatRippleModule,
-    MatSortModule,
-    MatPaginatorModule,
-    CdkTableModule,
-    DragDropModule,
-    OverlayModule,
-    PortalModule,
-    MatSelectModule,
-    MatCheckboxModule,
-    MatInputModule,
-    MatRadioModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatTabsModule,
-    MatTableModule,
-    MatSlideToggleModule,
-    MatButtonModule,
-    MatTooltipModule,
-    MatProgressBarModule,
-    LayoutModule,
-    MatProgressSpinnerModule
-  ]
+    imports: [
+        MatFormFieldModule,
+        MatRippleModule,
+        MatDialogModule,
+        MatSortModule,
+        MatPaginatorModule,
+        CdkTableModule,
+        DragDropModule,
+        PortalModule,
+        CommonModule,
+        MatSelectModule,
+        MatCheckboxModule,
+        MatInputModule,
+        MatRadioModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatTabsModule,
+        MatButtonModule,
+        MatTableModule,
+        MatSlideToggleModule,
+        OverlayModule,
+        MatTooltipModule,
+        MatProgressBarModule,
+        MatMenuModule,
+        MatProgressSpinnerModule
+    ],
+    declarations: [],
+    providers: [
+        { provide: MAT_RIPPLE_GLOBAL_OPTIONS, useValue: globalRippleConfig }
+    ],
+    exports: [
+        MatFormFieldModule,
+        PlatformModule,
+        MatRippleModule,
+        MatSortModule,
+        MatPaginatorModule,
+        CdkTableModule,
+        DragDropModule,
+        OverlayModule,
+        PortalModule,
+        MatSelectModule,
+        MatCheckboxModule,
+        MatInputModule,
+        MatRadioModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatTabsModule,
+        MatTableModule,
+        MatSlideToggleModule,
+        MatButtonModule,
+        MatTooltipModule,
+        MatProgressBarModule,
+        LayoutModule,
+        MatProgressSpinnerModule
+    ]
 })
 export class AppMaterialModule { }
